add 4 scripts of prototype for applying DRS4 sampling interval correction.
* The create_sampling_interval.py can create the counts table of coming pulses for all pixels to create sampling intervals table.
It should use when some LST-camera modules are changed.
* The ConvertCountsToInterval.py can merge sevelar counts tables made by create_sampling_interval.py and convert the counts to sampling intervals.
It should use when some LST-camera modules are changed.
* The charge_calculation.py is example script to apply sampling intervals (using ROOT).
It will be merged only the algorithm of applying sampling intervals to LSTR0Correction class.
The sampling interval correction should use to calibrate raw waveform.
* The merge_hist.py can merge the result files to one using ROOT.
It will be delete when the algorithm is added to LSTR0Correction class.