Update sensitivity.py
I was reporting an error due to the wrong shape of alpha in the calculation of the Li&Ma significance. 
Problem solved, together with improved output and readibility (prints)    

Fixed function >  calculate_sensitivity_lima
alpha was given only as an array of lenght equal to len(n_on), but it was not of the same shape (20,11,10).
Correcting the input  np.ones_like(final_gamma)) instead of np.ones(len(final_gamma)) solves the problem
                                                                        
Please experts check if the plots are correct or not. I did not change much, but for example the sensitivity seems not correct (was not correct also before).

Thank you
Luca