fixed calculation of noise in dim pixels parameter
Unnecessary square root was added so the noise in dim pixels was defined as standard deviation.
While later on it is used as a number of NSB photons from which poissonian number is selected (that adds this square root in the noise).
Noise in the brigth pixels was properly defined (without square root)