disp calculation
Simplified calculation of disp. Also, disp_norm slightly changes meaning. Before it was just "dist" (true distance cog to source), now it is distance along reconstructed axis from cog to the point in the axis closest to the source. This is more what we are looking for, so it should be better (difference is likely minimal).