Update muon_analysis.py
With these changes, the fit_muon function now fits a ring correctly even under high noise conditions. This was previously not the case. The changes made are:

the min. number of pixel neighbors to survive the tailcuts cleaning is now set to 2 (previously was set to default, which is 0).
If no tailcuts are indicated as an input, the function will select appropiate tailcuts based on the noise of the image.
Also, in the analyze_muon_event function, when calling the fit_muon function in line 231, the 'tailcuts' input has been removed. This way, when using analyze_muon_event, the tailcuts will be adapted to the image's noise.