Fixed long-standing BUG which made that the info in the DVR_settings file produced by the script contained only the information of the last subrun.
Improvements in logging

Now by default only 10 files (subruns) of the input set, distributed uniformly
along the run, are actually processed to determine the DVR settings. This is
to speed up the process (since the settings are determined for the whole run,
we do not need to check all subruns).