Now configuration is fully made through the config file, no command-line options
The interpolation can be switched on for the three reconstructions
independently (energy, gammaness, direction)

By default the correction is applied. If no MC DL1 training
directory is provided, the path will be built from the path
of the RF models. If directory does not exist the option is
simply deactivated