Remove unused "tailcut" settings
To avoid confusion, we set the "tailcut" setting in the json file (which is not used later) to {} (empty), rather than the default 8,4 (picture, boundary). The default appeared in the attrs of the resulting DL1 file, which may lead to confusion. The actually used settings are those in tailcuts_clean_with_pedestal_threshold