Keep all peak times in DL1a
Currently the output DL1 file keeps only the times of pixels that fulfill all cleaning conditions, except the time condition.
This change makes that all pixel times are kept. This is relevant e.g. if one wants to apply lower tailcuts to re-generate DL1b. Equivalently, if one wants to re-generate DL1b with a scaling (>1) of the pixel charges and the same tailcuts (because times will be missing  for the additional pixels which survive the cleaning charge-wise)