get_intensity_cut behavior
Until now, if finding of the intensity peak failed, a log message was shown telling so, but the returned value was 50 p.e. Besides, another message was shown indicating that 50 p.e. was an adequate cut for the run... So it was confusing.
This was just due to the behaviour of the max function when the second argument is nan.
Now in case the peak is not found the function returns nan. It should however happen rarely.